+++ db/src/guild.rs
@@ -3,12 +3,12 @@ use sea_orm::entity::prelude::*;
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
 #[sea_orm(table_name = "Guilds")]
 pub struct Model{
-    #[sea_orm(primary_key)]
-    pub guild_id: i32,
+    #[sea_orm(primary_key, auto_increment=false)]
+    pub guild_id: i64,
     pub is_compliant: bool,
-    pub snap_channel_id: Option<i32>,
-    pub warn_channel_id: Option<i32>,
-    pub moderation_role_id: Option<i32>
+    pub snap_channel_id: Option<i64>,
+    pub warn_channel_id: Option<i64>,
+    pub moderation_role_id: Option<i64>
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
